% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRbase-utilities.R
\name{all-subsets}
\alias{all-subsets}
\alias{all_subsets}
\alias{all_subsets0}
\alias{allSubsets}
\title{Create all subsets}
\usage{
all_subsets(x)

all_subsets0(x)

allSubsets(x, g.sep = "+")
}
\arguments{
\item{x}{Vector}

\item{g.sep}{Pick a value which is not in x}
}
\description{
Create all subsets of a vector
}
