% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{api-cell_}
\alias{api-cell_}
\alias{cell2entry_}
\alias{make_plevels_}
\alias{entry2cell_}
\alias{next_cell_}
\alias{next_cell2_}
\alias{next_cell_slice_}
\alias{slice2entry_}
\alias{cell2entry_perm_}
\alias{perm_cell_entries_}
\title{Low level table cell operations implemented in c++}
\usage{
cell2entry_(cell, dim)

make_plevels_(dim)

entry2cell_(entry, dim)

next_cell_(cell, dim)

next_cell2_(cell, dim)

next_cell_slice_(cell, dim, slice_marg)

slice2entry_(slice_cell, slice_marg, dim)

cell2entry_perm_(cell, dim, perm)

perm_cell_entries_(perm, dim)
}
\arguments{
\item{cell}{Vector giving the cell, e.g. c(1, 1, 2) in 3-way table.}

\item{dim}{Vector giving array dimension, eg c(2, 2, 2).}

\item{entry}{An entry in an array (a number indexing a vector).}

\item{slice_marg}{Vector giving the margin of a table, eg. c(2, 3)}

\item{slice_cell}{Vector giving the corresponding cell of marginal
table, e.g. c(1, 2)}

\item{perm}{Vector giving permutaion of array, eg. c(1, 3, 2).}
}
\description{
Corresponding R
   functions without the trailing underscore exist.
}
