% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{api-tabX_}
\alias{api-tabX_}
\alias{tab_perm_}
\alias{tab_expand_}
\alias{tab_align_}
\alias{tab_marg_}
\alias{tab_op_}
\alias{tab_add_}
\alias{tab_subt_}
\alias{tab_mult_}
\alias{tab_div_}
\alias{tab_div0_}
\alias{tab_equal_}
\alias{tab_list_mult_}
\alias{tab_list_add_}
\title{Table operations implemented in c++}
\usage{
tab_perm_(tab, perm)

tab_expand_(tab, aux, type = 0L)

tab_align_(tab1, tab2)

tab_marg_(tab, marg)

tab_op_(tab1, tab2, op = "*")

tab_add_(tab1, tab2)

tab_subt_(tab1, tab2)

tab_mult_(tab1, tab2)

tab_div_(tab1, tab2)

tab_div0_(tab1, tab2)

tab_equal_(tab1, tab2, eps = 1e-12)

tab_list_mult_(lst)

tab_list_add_(lst)
}
\arguments{
\item{tab}{Arrays with named dimnames (we call them
'named arrays').}

\item{perm}{A vector of indices or dimnames or a right hand sided
formula giving the desired permutation/margin.}

\item{aux}{Either a list with names and dimnames or a named array
from which such a list can be extracted.}

\item{type}{One of "none", "first" and "all".}

\item{tab1}{Arrays with named dimnames (we call them
'named arrays').}

\item{tab2}{Arrays with named dimnames (we call them
'named arrays').}

\item{marg}{A vector of indices or dimnames or a right hand sided
formula giving the desired permutation/margin.}

\item{op}{The operation to be carried out; "+", "-", "*", "/".}

\item{eps}{Criterion for checking equality of two arrays.}

\item{lst}{List of arrays.}
}
\description{
Table operations implemented in c++. Corresponding R
   functions without the trailing underscore exist.
}
