% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_array_zar_X.R
\name{api_ar_dist}
\alias{api_ar_dist}
\alias{ar_dist}
\title{Marginalize and condition in multidimensional array.}
\usage{
ar_dist(tab, marg = NULL, cond = NULL, normalize = TRUE)
}
\arguments{
\item{tab}{Multidimensional array with dimnames.}

\item{marg}{A specification of the desired margin; a character vector, a
numeric vector or a right hand sided formula.}

\item{cond}{A specification of what is conditioned on. Can take two forms:
Form one is a a character vector, a numeric vector or a right hand sided
formula. Form two is as a simple slice of the array, which is a list of
the form var1=value1, var2=value2 etc.}

\item{normalize}{Should the result be normalized to sum to 1.}
}
\description{
Marginalize and condition in a multidimensional array
    which is assumed to represent a discrete multivariate
    distribution.
}
\examples{
ar_dist(HairEyeColor)
}
