% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_array_zar_X.R
\name{api_ar_new}
\alias{api_ar_new}
\alias{ar_new}
\title{Create multidimensional arrays}
\usage{
ar_new(names, levels, values, normalize = "none", smooth = 0)
}
\arguments{
\item{names}{Names of variables defining table; a character vector
or a right hand sided formula.}

\item{levels}{1) a list with specification of the levels of the
factors in \code{names} or 2) a vector with number of levels of
the factors in \code{names}. See 'examples' below.}

\item{values}{values to go into the parray}

\item{normalize}{Either "none", "first" or "all". Should result be
normalized, see 'Details' below.}

\item{smooth}{Should values be smoothed, see 'Details' below.}
}
\description{
Alternative ways of creating arrays
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
