% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_array_zar_X.R
\name{api_ar_normalize}
\alias{api_ar_normalize}
\alias{ar_normalize}
\alias{ar_norm}
\title{Normalize an array}
\usage{
ar_normalize(tab, type = "none")

ar_norm(tab, type = "none")
}
\arguments{
\item{tab}{Arrays with named dimnames (we call them
'named arrays').}

\item{type}{One of "none", "first" and "all".}
}
\value{
An array
}
\description{
Normalize an array in various ways.
}
\examples{
ar_normalize(HairEyeColor, type="first")
ar_normalize(HairEyeColor, type="all")
}
