% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_array_tabX.R
\name{api_tab_slice}
\alias{api_tab_slice}
\alias{tab_slice}
\alias{tab_slice_prim}
\alias{tab_slice_mult}
\alias{tab_slice_to_entries}
\alias{tabSliceMult}
\alias{tabSlice}
\title{Array slices}
\usage{
tab_slice(
  tab,
  slice = NULL,
  margin = names(slice),
  drop = TRUE,
  as.array = FALSE
)

tab_slice_prim(tab, slice, drop = TRUE)

tab_slice_mult(tab, slice, val = 1, comp = 0)

tab_slice_to_entries(tab, slice, complement = FALSE)

tabSliceMult(tab, slice, val = 1, comp = 0)

tabSlice(
  tab,
  slice = NULL,
  margin = names(slice),
  drop = TRUE,
  as.array = FALSE
)
}
\arguments{
\item{tab}{An array with named dimnames.}

\item{slice}{A list defining the slice.}

\item{margin}{Names of variables in slice.}

\item{drop}{If TRUE then dimensions with only one level will be
dropped from the output.}

\item{as.array}{If the resulting array is one-dimensional the
result will by default be a vector with no dim attribute unless
as.array is TRUE.}

\item{val}{The values that entries in the slice will be multiplied
with.}

\item{comp}{The values that entries NOT in the slice will be
multiplied with.}

\item{complement}{If TRUE the complement of the entries are
returned.}
}
\description{
Functions for extracting slices of arrays
}
\examples{

x = HairEyeColor
s = list(Hair=c("Black", "Brown"), Eye=c("Brown", "Blue"))

s1 = tab_slice(x, slice=s); s1

tab_slice_to_entries(x, slice=s)
tab_slice_to_entries(x, slice=s, complement=TRUE)

## tab_slice_mult 
s2 = tab_slice_mult(x, slice=s); s2

sp = list(c(1,2), c(1,2), TRUE)
tab_slice_prim(x, slice=sp)
tab_slice(x, slice=s)
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
