% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_array_api_cpp.R
\name{array-api}
\alias{array-api}
\alias{tabAdd}
\alias{tabOp__}
\alias{tabEqual__}
\alias{tabPerm__}
\alias{tabAlign__}
\alias{tabExpand__}
\alias{tabListAdd__}
\alias{tabListMult__}
\alias{tabListAdd}
\alias{tabListMult}
\alias{tabExt}
\alias{tab_align_}
\alias{tab_expand_}
\alias{tab_marg_}
\alias{tab_perm_}
\alias{tabAlign}
\alias{tabDiv}
\alias{tabDiv0}
\alias{tabEqual}
\alias{tabExpand}
\alias{tabListMult}
\alias{tabListAdd}
\alias{tabMult}
\alias{tabSubt}
\alias{tabPerm}
\alias{tabMarg}
\alias{tabSum}
\alias{tabProd}
\title{Interface - operations on multidimensional arrays.}
\usage{
tabAdd(tab1, tab2)

tabAlign(tab1, tab2)

tabDiv(tab1, tab2)

tabDiv0(tab1, tab2)

tabEqual(tab1, tab2, eps = 1e-12)

tabExpand(tab, aux)

tabListMult(lst)

tabListAdd(lst)

tabMult(tab1, tab2)

tabSubt(tab1, tab2)

tabPerm(tab, perm)

tabMarg(tab, marg = NULL)

tabSum(...)

tabProd(...)
}
\arguments{
\item{eps}{Criterion for checking equality of two arrays.}

\item{tab, tab1, tab2, ...}{Arrays with named dimnames (we call them
'named arrays').}

\item{aux}{Either a list with names and dimnames or a named array
from which such a list can be extracted.}

\item{lst}{List of arrays.}

\item{perm, marg}{A vector of indices or dimnames or a right hand sided
formula giving the desired permutation/margin.}
}
\description{
Interface functions and minor extensions to cpp functions.
}
