% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array-cell}
\alias{array-cell}
\alias{make_plevels_}
\alias{cell2entry_}
\alias{cell2entry_prim_}
\alias{next_cell_}
\alias{next_cell_slice_}
\alias{slice2entry_}
\alias{get_cell_number_}
\alias{perm_cell_entries_}
\alias{entry2cell_prim_}
\alias{entry2cell_}
\title{table cell operations.}
\usage{
make_plevels_(dim)

cell2entry_(cell, dim)

cell2entry_prim_(cell, plevels)

next_cell_(cell, dim)

next_cell_slice_(cell, dim, slice_marg)

slice2entry_(slice_cell, slice_marg, dim)

get_cell_number_(cell, dim, perm)

perm_cell_entries_(perm, dim)

entry2cell_prim_(entry, plevels)

entry2cell_(entry, dim)
}
\arguments{
\item{dim}{Vector giving array dimension, eg c(2,2,2).}

\item{cell}{Vector giving the cell, eg c(1,1,2) in 3-way table.}

\item{plevels}{A vector used for fast computations of entries from cells; generated by 'make_plevels_'.}

\item{slice_marg}{Vector giving the margin of a table, eg. c(2,3)}

\item{slice_cell}{Vector giving the corresponding cell of marginal table, eg. c(1,2)}

\item{perm}{Vector giving permutaion of array, eg. c(1,3,2).}

\item{entry}{An entry in an array (a number indexing a vector).}
}
\description{
low level table cell operations
}
