% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_array_cpp_ainterface.R
\name{array-interface}
\alias{array-interface}
\alias{tabAdd__}
\alias{tab_align_}
\alias{tab_expand_}
\alias{tab_marg_}
\alias{tab_perm_}
\alias{tabDiv__}
\alias{tabDiv0__}
\alias{tabMarg__}
\alias{tabMult__}
\alias{tabSubt__}
\alias{tabEqual}
\alias{tabAlign}
\alias{tabExpand}
\alias{tabPerm}
\alias{tabMarg}
\alias{tabAdd}
\alias{tabDiv}
\alias{tabDiv0}
\alias{tabMult}
\alias{tabSubt}
\alias{tabListMult}
\alias{tabListAdd}
\alias{tabSum}
\alias{tabProd}
\title{Interface - operations on multidimensional arrays.}
\usage{
tabAdd__(tab1, tab2)

tabEqual(tab1, tab2, eps = 1e-12)

tabAlign(tab1, tab2)

tabExpand(tab, aux)

tabPerm(tab, perm)

tabMarg(tab, marg = NULL)

tabAdd(tab1, tab2)

tabDiv(tab1, tab2)

tabDiv0(tab1, tab2)

tabMult(tab1, tab2)

tabSubt(tab1, tab2)

tabListMult(lst)

tabListAdd(lst)

tabSum(...)

tabProd(...)
}
\arguments{
\item{eps}{Criterion for checking equality of two arrays.}

\item{tab, tab1, tab2, ...}{Arrays with named dimnames (we call them
'named arrays').}

\item{aux}{Either a list with names and dimnames or a named array
from which such a list can be extracted.}

\item{perm}{A vector of indices or dimnames or a right hand sided
formula giving the desired permutiation.}

\item{marg}{A vector of indices or dimnames or a right hand sided
formula giving the desired marginal.}

\item{lst}{List of arrays.}
}
\description{
Interface functions and minor extensions to c++ functions.

tabSubt__ tabMult__ tabOp__
    tabEqual__ tabMarg__ tabPerm__ tabAlign__ tabExpand__
    tabListAdd__ tabListMult__ tabListAdd tabListMult tabExt
}
