% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_arDist.R
\name{array-normalize}
\alias{ar_normalize}
\alias{array-normalize}
\alias{tabNormalize}
\title{Normalize an array}
\usage{
tabNormalize(tab, type = "none")

ar_normalize(tab, type = "none")
}
\arguments{
\item{tab}{A multidimensional array}

\item{type}{Either "none", "first", or "all"}
}
\value{
An array
}
\description{
Normalize an array in various ways.
}
\examples{
ar_normalize( HairEyeColor, type="first")
ar_normalize( HairEyeColor, type="all")
}

