% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_array_cpp_properties.R
\name{array-properties}
\alias{array-properties}
\alias{is.named.array}
\alias{is_named_array_}
\alias{is_number_vector_}
\alias{is_dimnames_}
\title{Check if object is array}
\usage{
is.named.array(obj)

is_named_array_(obj)

is_number_vector_(obj)

is_dimnames_(obj)
}
\arguments{
\item{obj}{Some R object.}
}
\description{
Check if object is array (that it is a vector with a
    dim attribute) and that the object has dimnames and that
    dimnames are named.
}
\examples{

is.named.array( HairEyeColor )
is.named.array( matrix(1:4, nrow=2) )
is_named_array_( HairEyeColor )
is_named_array_( matrix(1:4, nrow=2) )
is_number_vector_(1:4)
is_number_vector_(list(1:4))

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
