% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_array_simulate.R
\name{array_simulate}
\alias{array_simulate}
\alias{simulate.table}
\alias{simulate.xtabs}
\alias{simulate.array}
\alias{simulateArray}
\title{Simulate data from array.}
\usage{
\method{simulate}{table}(object, nsim = 1, seed = NULL, margin, value.margin, ...)

\method{simulate}{xtabs}(object, nsim = 1, seed = NULL, margin, value.margin, ...)

\method{simulate}{array}(object, nsim = 1, seed = NULL, margin, value.margin, ...)

simulateArray(x, nsim = 1, margin, value.margin, seed = NULL)
}
\arguments{
\item{nsim}{Number of cases to simulate.}

\item{seed}{Seed to be used for random number generation.}

\item{margin, value.margin}{Specification of slice of array to
simulate from.}

\item{...}{Additional arguments, currently not used.}

\item{x, object}{An array.}
}
\value{
A matrix.
}
\description{
Simulate data (slice of) an array: Simulate n
observations from the array x conditional on the variables in
margin (a vector of indices) takes values given by margin.value
}
\details{
A multidimensional table of numbers is represented by a
multidimensional array, so we can use the terms 'table' and
'array' interchangeably. In this context, 'table' refers
specifically to numerical data structured in multiple
dimensions, similar to how arrays are used in programming. An
alternative representation of a multidimensional table would be
as a dataframe.
}
\note{
The current implementation is fragile in the sense that it is
not checked that the input argument \code{x} is an array.
}
\examples{
## 2x2 array
x <- tab_new(c("a", "b"), levels=c(2, 2), values=1:4)

## Simulate from entire array
s <- simulate_array(x, 1000)
xtabs(~., as.data.frame(s))

## Simulate from slice defined by that dimension 1 is fixed at level 2
s <-simulate_array(x, 1000, margin=1, value.margin=2)
xtabs(~., as.data.frame(s))

## 2 x 2 x 2 array
x <- tab_new(c("a", "b", "c"), levels=c(2, 2, 2), values=1:8)
## Simulate from entire array
s <-simulate_array(x, 36000)
xtabs(~., as.data.frame(s))

## Simulate from slice defined by that dimension 3 is fixed at level 1
s <-simulate_array(x, 10000, 3, 1)
xtabs(~., as.data.frame(s))

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
