\name{arraypack-internal}

%% Array operations in C
%% ---------------------
%% Examples of the usage of these functions are given in
%% vignette
%% \alias{aperm2}
%% \alias{cell2entry}
%% \alias{entry2cell}
%% \alias{factGrid}
%% \alias{getCellNumberC}
%% \alias{slice2entry}
%% \alias{nextCell}
%% \alias{nextCellSlice}
%% \alias{permuteCellEntries}


\alias{allSubsets__}
\alias{allSubsets0__}


\alias{cell2entry2_cpp}
\alias{cell2entry_cpp}
\alias{getCellNumberPrim_cpp}
\alias{getCellNumber_cpp}
\alias{nextCellSlicePrim_cpp}
\alias{nextCellSlice_cpp}
\alias{nextCell_cpp}
\alias{permuteCellEntries_cpp}
\alias{slice2entry_cpp}



%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}

%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}

%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}

%\alias{}
%\alias{}

%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}



\title{Internal arraypack functions}
\description{
  Internal functions called by other functions.
}
\keyword{internal}

