% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{chestSim}
\alias{chestSim}
\alias{chestSim500}
\alias{chestSim1000}
\alias{chestSim10000}
\alias{chestSim50000}
\alias{chestSim100000}
\title{Simulated data from the Chest Clinic example}
\format{A data frame with 500 observations on the following 8 variables.
  \describe{
  \item{\code{asia}}{a factor with levels \code{yes} \code{no}}
  \item{\code{tub}}{a factor with levels \code{yes} \code{no}}
  \item{\code{smoke}}{a factor with levels \code{yes} \code{no}}
  \item{\code{lung}}{a factor with levels \code{yes} \code{no}}
  \item{\code{bronc}}{a factor with levels \code{yes} \code{no}}
  \item{\code{either}}{a factor with levels \code{yes} \code{no}}
  \item{\code{xray}}{a factor with levels \code{yes} \code{no}}
  \item{\code{dysp}}{a factor with levels \code{yes} \code{no}}
}}
\usage{
chestSim500
}
\description{
Simulated data from the Chest Clinic example (also known as the Asia
example) from Lauritzen and Spiegelhalter, 1988.
}
\examples{

data(chestSim500)
## maybe str(chestSim500) ; plot(chestSim500) ...

}
\references{
Lauritzen and Spiegelhalter (1988) Local Computations
    with Probabilities on Graphical Structures and their
    Application to Expert Systems (with
    Discussion). J. Roy. Stat. Soc. 50, p. 157-224.
}
\keyword{datasets}
