% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{compareModels}
\alias{compareModels}
\title{Generic function for model comparison}
\usage{
compareModels(object, object2, ...)
}
\arguments{
\item{object, object2}{Model objects}

\item{\dots}{Additional arguments}
}
\value{
The value returned depends on the class of the first
    argument.
}
\description{
\code{compareModels} is a generic functions which
    invoke particular methods which depend on the class of the
    first argument
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
