% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{data-dietox}
\alias{data-dietox}
\alias{dietox}
\title{Growth curves of pigs in a 3x3 factorial experiment}
\format{This data frame contains the following columns: Weight,
    Feed, Time, Pig, Evit, Cu, Litter.}
\source{
Lauridsen, C., Højsgaard, S., Sørensen, M.T. C. (1999)
    Influence of Dietary Rapeseed Oli, Vitamin E, and Copper on
    Performance and Antioxidant and Oxidative Status of
    Pigs. J. Anim. Sci.77:906-916
}
\usage{
data(dietox)
}
\description{
The \code{dietox} data frame has 861 rows and 7 columns.
}
\examples{

data(dietox)

}
\keyword{datasets}
