% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_grbase.R
\docType{data}
\name{data-personality}
\alias{data-personality}
\alias{personality}
\title{Personality traits}
\format{
This dataframe has recordings on the following 32
variables: distant, talkatv, carelss, hardwrk, anxious,
agreebl, tense, kind, opposng, relaxed, disorgn, outgoin,
approvn, shy, discipl, harsh, persevr, friendl, worryin,
respnsi, contrar, sociabl, lazy, coopera, quiet, organiz,
criticl, lax, laidbck, withdrw, givinup, easygon
}
\usage{
data(personality)
}
\description{
The \code{peronality} dataframe has 240 rows and 32 columns
}
\examples{
data(personality)
str(personality)
}
\references{
Origin unclear
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\concept{dataset}
\keyword{datasets}
