% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_grbase.R
\docType{data}
\name{data-rats}
\alias{data-rats}
\alias{rats}
\title{Weightloss of rats}
\format{
A dataframe with 4 variables. Sex: "M" (male), "F" (female). Drug:
"D1", "D2", "D3" (three types). W1 weightloss, week one. W2 weightloss, week
2.
}
\usage{
data(rats)
}
\description{
An artificial dataset. 24 rats (12 female, 12 male) have been randomized to
use one of three drugs (products for loosing weight). The weightloss for
each rat is noted after one and two weeks.
}
\references{
Morrison, D.F. (1976). Multivariate Statistical Methods.
McGraw-Hill, USA.

Edwards, D. (1995). Introduction to Graphical Modelling, Springer-Verlag.
New York.
}
\concept{dataset}
\keyword{datasets}
