% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyOLD_grbase.R
\name{data_handling}
\alias{data_handling}
\alias{valueLabels}
\alias{valueLabels.default}
\alias{varNames}
\alias{varNames.default}
\alias{nLevels}
\alias{nLevels.default}
\title{Utilities for data handling}
\usage{
valueLabels(x)

\method{valueLabels}{default}(x)

varNames(x)

\method{varNames}{default}(x)

nLevels(x)

\method{nLevels}{default}(x)
}
\arguments{
\item{x}{Data, typically a dataframe.}
}
\description{
Utilities for data handling
}
