\name{dataOrNULL-class}
\docType{class}
\alias{dataOrNULL-class}

\title{Class "dataOrNULL"}
\description{ Objects of class \code{gmData-class} may or may not
  contain data. }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
%\section{Methods}{
%No methods defined with class "dataOrNULL" in the signature.
%}
%\references{ ~put references to the literature/web site here ~ }
\author{ Claus Dethlefsen and Søren Højsgaard }
%\note{ ~~further notes~~ }

% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 \code{\link{gmData-class}} 
}
%\examples{
%##---- Should be DIRECTLY executable !! ----
%}
\keyword{classes}
