% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_downstream-depends.R
\name{downstream-aliases}
\alias{downstream-aliases}
\alias{tabAdd__}
\alias{tabDiv__}
\alias{tabDiv0__}
\alias{tabMarg__}
\alias{tabMult__}
\alias{tabSubt__}
\alias{topoSort}
\alias{topoSort.default}
\alias{topoSortMAT}
\alias{subsetof}
\alias{removeRedundant}
\alias{mcsmarked}
\alias{jTree}
\alias{jTree.default}
\alias{getCliques}
\alias{maxCliqueMAT}
\alias{combnPrim}
\alias{mcsmarkedMAT}
\alias{nextCell}
\alias{ell}
\alias{ellK}
\alias{isin}
\alias{is.TUG}
\alias{is.DAG}
\alias{graphNEL2adjMAT}
\alias{glist2adjMAT}
\alias{is_subsetof_}
\alias{get_superset_}
\alias{get_subset_}
\alias{is.UG}
\alias{is.DG}
\alias{graphNEL2M}
\alias{M2graphNEL}
\alias{tab}
\title{Downstream aliases}
\description{
Downstream aliases for other graphical modelling
    packages.  Will be deprecated in due course.
}
