% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_gmwr-depends.R
\name{gmwr-book}
\alias{gmwr-book}
\alias{as.adjMAT}
\title{Functions from Graphical Modelling with R book}
\usage{
as.adjMAT(object, result = "matrix")
}
\arguments{
\item{object}{An object to be coerced.}

\item{result}{The format to be coerced to.}
}
\description{
Functions that must be retained to make code from
    gmwr-book work
}
