% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_gmwr_depends.R
\name{gmwr_book}
\alias{gmwr_book}
\alias{as.adjMAT}
\title{Functions from Graphical Modelling with R book}
\usage{
as.adjMAT(object)
}
\arguments{
\item{object}{An object to be coerced.}
}
\description{
Functions that must be retained to make code from
gmwr-book work
}
