% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce-api.R
\name{graph-coerce-api}
\alias{graph-coerce-api}
\alias{gn2dm_}
\alias{gn2sm_}
\alias{gn2ig_}
\alias{dm2gn_}
\alias{dm2sm_}
\alias{dm2ig_}
\alias{sm2gn_}
\alias{sm2dm_}
\alias{sm2ig_}
\alias{ig2gn_}
\alias{ig2dm_}
\alias{ig2sm_}
\alias{as_}
\alias{as_.graphNEL}
\alias{as_.matrix}
\alias{as_.dgCMatrix}
\alias{as_.igraph}
\title{API for coercing graph representations}
\usage{
gn2dm_(xx)

gn2sm_(xx)

gn2ig_(xx)

dm2gn_(xx)

dm2sm_(xx)

dm2ig_(xx)

sm2gn_(xx)

sm2dm_(xx)

sm2ig_(xx)

ig2gn_(xx)

ig2dm_(xx)

ig2sm_(xx)

as_(object, Class)

\method{as_}{graphNEL}(object, Class)

\method{as_}{matrix}(object, Class)

\method{as_}{dgCMatrix}(object, Class)

\method{as_}{igraph}(object, Class)
}
\arguments{
\item{xx}{An object representing a graph}

\item{object}{A graph object}

\item{Class}{The desired output class}
}
\description{
API for coercing graph representations.
}
\details{
No checking is made. In the function the following names are used:
"ig": "igraph";
"gn": "graphNEL";
"sm": "dgCMatrix" (sparse matrix);
"dm": "matrix" (dense matrix)

The 'as_' functions are experimental and may well change
    in the future.
}
\examples{
## FIXME examples coerce graph missing


g <- ug(~ 1:2+2:3)
as_(g, "matrix")
as_(g, "dgCMatrix")
as_(g, "igraph")

if (require(microbenchmark)){
microbenchmark(
  as_(g, "matrix"),
  coerceGraph(g, "matrix"),
  as(g, "matrix"),

  as_(g, "dgCMatrix"),
  coerceGraph(g, "dgCMatrix"),
  as(g, "dgCMatrix")
)
}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
## @seealso \code{\link{ug}}, \code{\link{dag}}
}
\keyword{utilities}
