% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce-list.R
\name{M2adjList}
\alias{M2adjList}
\alias{adjList2dgCMatrix_}
\alias{adjList2matrix_}
\alias{adjList2ftList_}
\alias{adjList2ftM_}
\alias{adjList2tfList_}
\alias{adjList2tfM_}
\alias{adl2dm}
\alias{adl2gn}
\alias{adl2ig}
\alias{adl2sm}
\alias{dagList2dgCMatrix_}
\alias{dagList2matrix_}
\alias{dgCMatrix2matrix_}
\alias{dgl2dm}
\alias{dgl2gn}
\alias{dgl2ig}
\alias{dgl2sm}
\alias{matrix2dgCMatrix_}
\alias{ugList2dgCMatrix_}
\alias{ugList2matrix_}
\alias{ugl2dm}
\alias{ugl2gn}
\alias{ugl2ig}
\alias{ugl2sm}
\alias{M2ugList}
\alias{M2dagList}
\alias{ugList2graphNEL}
\alias{ugList2M}
\alias{dagList2graphNEL}
\alias{dagList2M}
\alias{adjList2M}
\alias{ugl2gn_}
\alias{ugl2dm_}
\alias{ugl2sm_}
\alias{ugl2ig_}
\alias{dgl2gn_}
\alias{dgl2dm_}
\alias{dgl2sm_}
\alias{dgl2ig_}
\alias{adl2gn}
\alias{adl2dm}
\alias{adl2sm}
\alias{adl2ig}
\title{Coercion of graphs represented as lists}
\usage{
M2adjList(amat)

M2ugList(amat)

M2dagList(amat)

ugList2graphNEL(glist, vn = NULL)

ugList2M(glist, vn = NULL, result = "matrix")

dagList2graphNEL(glist, vn = NULL)

dagList2M(glist, vn = NULL, result = "matrix")

adjList2M(alist, result = "matrix")

ugl2gn_(zz)

ugl2dm_(zz)

ugl2sm_(zz)

ugl2ig_(zz)

dgl2gn_(zz)

dgl2dm_(zz)

dgl2sm_(zz)

dgl2ig_(zz)

adl2gn(zz)

adl2dm(zz)

adl2sm(zz)

adl2ig(zz)
}
\arguments{
\item{amat}{Adjacency matrix (dense or sparse dgCMatrix).}

\item{glist}{A list of generators where a generator is a character
vector. If interpreted as generators of an undirected graph, a
generator is a complete set of vertices in the graph. If
interpreted as generators of a dag, a generator (v1,...,vn)
means that there will be arrows from v2,...,vn to v1.}

\item{vn}{The names of the vertices in the graphs. These will be
the row and column names of the matrix}

\item{result}{FIXME}

\item{alist}{An adjacency list.}

\item{zz}{An object representing a graph.}
}
\description{
FIXME
}
\examples{

## Sparse and dense adjacency matrices converted to adjacency list 
g1 <- ug(~a:b + b:c + c:d, result="matrix")
g2 <- ug(~a:b + b:c + c:d, result="dgCMatrix")
M2adjList( g1 )
## M2adjList( g2 ) ## FIXME FAILS for sparse matrix

## Sparse and dense adjacency matrices converted to cliques
M2ugList( g1 )
## M2ugList( g2 ) ## FIXME Is there an issue here??
## Sparse and dense adjacency matrices converted to cliques
M2dagList( g1 )
## M2dagList( g2 ) ## Fails for sparse matrix

}
