% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce.R
\name{graph-coerce}
\alias{coerceGraph}
\alias{coerceGraph.dgCMatrix}
\alias{coerceGraph.graphNEL}
\alias{coerceGraph.igraph}
\alias{coerceGraph.matrix}
\alias{graph-coerce}
\title{Graph coercion}
\usage{
coerceGraph(object, result)

\method{coerceGraph}{graphNEL}(object, result)

\method{coerceGraph}{matrix}(object, result)

\method{coerceGraph}{dgCMatrix}(object, result)

\method{coerceGraph}{igraph}(object, result)
}
\arguments{
\item{object}{A graph object}

\item{result}{The desired output type}
}
\description{
Methods for changing graph representations
}

