% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{ug2dag}
\alias{ug2dag}
\title{Coerce between undirected and directed graphs when possible}
\usage{
ug2dag(gn)
}
\arguments{
\item{gn}{A graphNEL object or an object that can be converted to a
graphNEL object.}
}
\description{
An undirected graph G can be converted to a dag if G
is chordal. A dag D can be converted to an undirected graph if
D can be moralized without adding edges.
}
