% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce.R
\name{graph_coerce}
\alias{graph_coerce}
\alias{coerceGraph}
\alias{graph_as}
\title{Graph coercion}
\usage{
coerceGraph(object, class)

graph_as(object, outtype, intype = NULL)
}
\arguments{
\item{object}{A graph object}

\item{class}{The desired output class}

\item{outtype}{The desired output outtype}

\item{intype}{The desired output outtype (only relevant if object is a list)}
}
\description{
Methods for changing graph representations
}
\details{
coerceGraph is used in the book "Graphical models with R".
A more generic approach is as().

g1 <- ug(~a:b+b:c)
as(g1, "igraph")
as(g1, "matrix")
as(g1, "Matrix")
as(g1, "dgCMatrix")

graph_as(g1, "ugList") 

l1 <- list(c("a" ,"b"), c("b", "c"))

graph_as(l1, "graphNEL", "ugList")
}
