% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce-xxx2yyy.R
\name{graph_coerce_api}
\alias{graph_coerce_api}
\alias{gn2dm_}
\alias{gn2sm_}
\alias{gn2ig_}
\alias{dm2gn_}
\alias{dm2sm_}
\alias{dm2ig_}
\alias{sm2gn_}
\alias{sm2dm_}
\alias{sm2ig_}
\alias{ig2gn_}
\alias{ig2dm_}
\alias{ig2sm_}
\alias{xm2gn_}
\alias{xm2ig_}
\alias{xm2dm_}
\alias{xm2sm_}
\alias{xm2xm_}
\alias{gn2xm_}
\alias{gn2ftM_}
\alias{gn2tfM_}
\title{API for coercing graph representations}
\usage{
gn2dm_(object)

gn2sm_(object)

gn2ig_(object)

dm2gn_(object)

dm2sm_(object)

dm2ig_(object)

sm2gn_(object)

sm2dm_(object)

sm2ig_(object)

ig2gn_(object)

ig2dm_(object)

ig2sm_(object)

xm2gn_(object)

xm2ig_(object)

xm2dm_(object)

xm2sm_(object)

xm2xm_(object, result = "matrix")

gn2xm_(object, result = "matrix")

gn2ftM_(object)

gn2tfM_(object)
}
\arguments{
\item{object}{An object representing a graph}

\item{result}{Either 'matrix' (dense) or 'dgCMatrix' (sparse, can
be abbreviated to 'Matrix').}
}
\description{
API for coercing graph representations.
}
\details{
No checking is made. In the function the following names are used:
"ig": "igraph";
"gn": "graphNEL";
"sm": "dgCMatrix" (sparse matrix);
"dm": "matrix" (dense matrix)
}
\seealso{
\code{\link{ug}}, \code{\link{dag}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
