% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_iplot.R
\name{graph_iplot}
\alias{graph_iplot}
\alias{iplot}
\alias{iplot.igraph}
\title{Function for plotting graphs using the 'igraph' package.}
\usage{
iplot(x, ...)

\method{iplot}{igraph}(x, ...)
}
\arguments{
\item{x}{A graph object to be plotted.}

\item{\dots}{Additional arguments}
}
\description{
Generic function for plotting graphs using the
'igraph' package and a plot method for graphNEL objects.
}
\examples{

UG <- ug(~a:b+b:c:d)
iplot(UG)

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{graphics}
