\name{mathmark}
\alias{mathmark}
\alias{math}

\title{Mathematics marks for students}
\usage{data(mathmark)}
\description{
The \code{mathmark} data frame has 88 rows and 5 columns.
}
\format{
  This data frame contains the following columns: mechanics, vectors,
  algebra, analysis, statistics.
  }

\references{ David Edwards, An Introduction to Graphical Modelling,
  Second Edition, Springer Verlag, 2000}
\author{Søren Højsgaard, sorenh@agrsci.dk}
\examples{
data(mathmark)
}
\keyword{datasets}

