% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_array-operations-bristol.R
\name{old-array-operations}
\alias{old-array-operations}
\alias{tableDiv}
\alias{tableGetSliceIndex}
\alias{tableMargin}
\alias{tableMult}
\alias{tableOp}
\alias{tableOp2}
\alias{tablePerm}
\alias{tableSetSliceValue}
\alias{tableSlice}
\alias{tableSlicePrim}
\title{Array operations (2007)}
\usage{
tablePerm(tab, perm, resize = TRUE, keep.class = FALSE)

tableMult(tab1, tab2)

tableDiv(tab1, tab2)

tableOp(tab1, tab2, op = "*")

tableOp2(tab1, tab2, op = `*`, restore = FALSE)

tableSlice(tab, margin, level, impose)

tableSlicePrim(tab, mar.idx, lev.idx)

tableMargin(tab, margin, keep.class = FALSE)

tableGetSliceIndex(tab, margin, level, complement = FALSE)

tableSetSliceValue(tab, margin, level, complement = FALSE, value = 0)
}
\arguments{
\item{tab, tab1, tab2}{Arrays with named dimnames.}

\item{perm}{A permutation; either indices or names.}

\item{resize}{A flag indicating whether the vector should be
resized as well as having its elements reordered (default
TRUE).}

\item{keep.class}{Obsolete argument.}

\item{op}{The operation; choices are \code{"*"}, \code{"/"}, \code{"+"}, \code{"-"}.}

\item{restore}{Not so clear anymore.}

\item{margin}{Index or name of margin.}

\item{level}{Corresponding level of margin.}

\item{impose}{Value to be imposed.}

\item{mar.idx}{Index of margin}

\item{lev.idx}{Index of level}

\item{complement}{Should values be set for the complement?}

\item{value}{Which value should be set}
}
\description{
Array operations; created to facilitate the gRain package in
    2007. Now largely replaceable by other faster functions implemented in
    Rcpp.
}

