% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_array-simulate-R.R
\name{simulateArray}
\alias{simulateArray}
\title{Simulate data from array.}
\usage{
simulateArray(x, nsim = 1, margin, value.margin)
}
\arguments{
\item{x}{An array}

\item{nsim}{Number of cases to simulate}

\item{margin, value.margin}{Specification of slice of array to
simulate from}
}
\value{
A matrix
}
\description{
Simulate data (slice of) an array.
}
\note{
The current implementation is fragile in the sense that it is
    not checked that the input argument \code{x} is an array.
}
\examples{


## 2x2 array
x <- parray(c("a","b"), levels=c(2,2), values=1:4)

## Simulate from entire array
s <-simulateArray(x,1000)
xtabs(~., as.data.frame(s))

## Simulate from slice defined by that dimension 1 is fixed at level 2
s <-simulateArray(x,6000,1,2)
xtabs(~., as.data.frame(s))

## 2x2x2 array
x <- parray(c("a","b","c"), levels=c(2,2,2), values=1:8)

## Simulate from entire array
s <-simulateArray(x,36000)
xtabs(~., as.data.frame(s))

## Simulate from slice defined by that dimension 3 is fixed at level 1
s <-simulateArray(x,10000,3,1)
xtabs(~., as.data.frame(s))


}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}

