% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_setops-c.R
\name{all_subsets}
\alias{all_subsets}
\alias{all_subsets0}
\title{Create all subsets}
\usage{
all_subsets(x)

all_subsets0(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Create all subsets of a vector
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
