% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_array_tabX.R
\name{api-tabX}
\alias{api-tabX}
\alias{tabAdd}
\alias{tabAlign}
\alias{tabDiv}
\alias{tabDiv0}
\alias{tabEqual}
\alias{tabExpand}
\alias{tabMult}
\alias{tabSubt}
\alias{tabListMult}
\alias{tabListAdd}
\alias{tabPerm}
\alias{tabMarg}
\alias{tabSum}
\alias{tabProd}
\alias{tabNormalize}
\title{Interface - operations on multidimensional arrays.}
\usage{
tabAdd(tab1, tab2)

tabAlign(tab1, tab2)

tabDiv(tab1, tab2)

tabDiv0(tab1, tab2)

tabEqual(tab1, tab2, eps = 1e-12)

tabExpand(tab, aux)

tabMult(tab1, tab2)

tabSubt(tab1, tab2)

tabListMult(lst)

tabListAdd(lst)

tabPerm(tab, perm)

tabMarg(tab, marg = NULL)

tabSum(tab, ...)

tabProd(tab, ...)

tabNormalize(tab, type = "none")
}
\arguments{
\item{eps}{Criterion for checking equality of two arrays.}

\item{tab, tab1, tab2, ...}{Arrays with named dimnames (we call them
'named arrays').}

\item{aux}{Either a list with names and dimnames or a named array
from which such a list can be extracted.}

\item{lst}{List of arrays.}

\item{perm, marg}{A vector of indices or dimnames or a right hand sided
formula giving the desired permutation/margin.}

\item{type}{One of "none", "first" and "all".}
}
\description{
Interface functions and minor extensions to cpp functions.
}
