% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array-cell}
\alias{array-cell}
\alias{cell2entry_}
\alias{next_cell_}
\alias{next_cell_slice_}
\alias{slice2entry_}
\alias{get_cell_number_}
\alias{perm_cell_entries_}
\title{table cell operations.}
\usage{
cell2entry_(cell, dim)

next_cell_(cell, dim)

next_cell_slice_(cell, dim, slice_set)

slice2entry_(slice_cell, slice_set, dim)

get_cell_number_(cell, dim, perm)

perm_cell_entries_(perm, dim)
}
\arguments{
\item{cell}{Vector giving the cell, eg c(1,1,2) in 3-way table.}

\item{dim}{Vector giving array dimension, eg c(2,2,2).}

\item{slice_set}{Vector giving the margin of a table, eg. c(2,3)}

\item{slice_cell}{Vector giving the corresponding cell of marginal table, eg. c(1,2)}

\item{perm}{Vector giving permutaion of array, eg. c(1,3,2).}
}
\description{
low level table cell operations
}
