% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_arDist.R
\name{array-normalize}
\alias{arnormalize}
\alias{array-normalize}
\title{Normalize an array}
\usage{
arnormalize(tab, type = "none")
}
\arguments{
\item{tab}{A multidimensional array}

\item{type}{Either "none", "first", or "all"}
}
\value{
An array
}
\description{
Normalize an array in various ways.
}
\examples{
arnormalize( HairEyeColor, type="first")
arnormalize( HairEyeColor, type="all")
}

