\name{cadcomplete}
\alias{cadcomplete}
\alias{cadincomplete}
\docType{data}
\title{Coronary artery disease data}
\description{
Coronary artery disease data
}
\usage{data(cadcomplete)}
\format{
  A data frame with 236 observations on the following 14 variables: Sex,
  AngPec, AMI, QWave, QWavecode, STcode, STchange, SuffHeartF,
  Hypertrophi, Hyperchol, Smoker, Inherit, Heartfail, CAD
  }
}
\examples{
data(cadcomplete)
}
\keyword{datasets}

