% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{dag2chol}
\alias{dag2chol}
\title{Create regression matrix matrix from DAG
A DAG can be represented as a triangular matrix of regression coefficients.}
\usage{
dag2chol(object)
}
\arguments{
\item{object}{A graph, either a graphNEL or an igraph object.}
}
\description{
Create regression matrix matrix from DAG
A DAG can be represented as a triangular matrix of regression coefficients.
}
\examples{

g <- dag(~x2|x1 + x3|x1:x2 + x4|x3)
gi <- as(g, "igraph")
dag2chol(g)
dag2chol(gi)

}
