% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_downstream_depends.R
\name{downstream-aliases}
\alias{downstream-aliases}
\alias{ell}
\alias{ellK}
\title{Downstream aliases}
\description{
Downstream aliases for other graphical modelling
packages.  Will be deprecated in due course.
}
