% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{downstream-aliases}
\alias{downstream-aliases}
\alias{tabAdd__}
\alias{tabDiv__}
\alias{tabDiv0__}
\alias{tabMarg__}
\alias{tabMult__}
\alias{tabSubt__}
\title{Downstream aliases}
\description{
Downstream aliases for other graphical modelling
    packages.  May change without further notice.
}
