% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce-xxx2yyy.R
\name{graph-coerce-api}
\alias{graph-coerce-api}
\alias{g_gn2dm_}
\alias{g_gn2sm_}
\alias{g_gn2ig_}
\alias{g_dm2gn_}
\alias{g_dm2sm_}
\alias{g_dm2ig_}
\alias{g_sm2gn_}
\alias{g_sm2dm_}
\alias{g_sm2ig_}
\alias{g_ig2gn_}
\alias{g_ig2dm_}
\alias{g_ig2sm_}
\alias{g_xm2gn_}
\alias{g_xm2ig_}
\alias{g_xm2dm_}
\alias{g_xm2sm_}
\alias{g_xm2xm_}
\alias{g_gn2xm_}
\alias{g_gn2ftM_}
\alias{g_gn2tfM_}
\title{API for coercing graph representations}
\usage{
g_gn2dm_(object)

g_gn2sm_(object)

g_gn2ig_(object)

g_dm2gn_(object)

g_dm2sm_(object)

g_dm2ig_(object)

g_sm2gn_(object)

g_sm2dm_(object)

g_sm2ig_(object)

g_ig2gn_(object)

g_ig2dm_(object)

g_ig2sm_(object)

g_xm2gn_(object)

g_xm2ig_(object)

g_xm2dm_(object)

g_xm2sm_(object)

g_xm2xm_(object, result = "matrix")

g_gn2xm_(object, result = "matrix")

g_gn2ftM_(object)

g_gn2tfM_(object)
}
\arguments{
\item{object}{An object representing a graph}

\item{result}{Either 'matrix' (dense) or 'dgCMatrix' (sparse, can
be abbreviated to 'Matrix').}
}
\description{
API for coercing graph representations.
}
\details{
No checking is made. In the function the following names are used:
* "ig": "igraph";
* "gn": "graphNEL";
* "sm": "dgCMatrix" (sparse matrix);
* "dm": "matrix" (dense matrix)
}
\seealso{
\code{\link{ug}}, \code{\link{dag}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
