% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coerce.R
\name{graph-coerce}
\alias{graph-coerce}
\alias{coerceGraph}
\title{Graph coercion}
\usage{
coerceGraph(object, Class)
}
\arguments{
\item{object}{A graph object}

\item{Class}{The desired output class}
}
\description{
Methods for changing graph representations
}
\details{
coerceGraph is used in the book "Graphical models with R".
A more generic approach is as().
}
