% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-querygraph.R
\name{graph-query}
\alias{graph-query}
\alias{querygraph}
\alias{ancestors}
\alias{ancestralSet}
\alias{parents}
\alias{children}
\alias{closure}
\alias{simplicialNodes}
\alias{ancestralGraph}
\alias{is.complete}
\alias{is.decomposition}
\alias{is.simplicial}
\title{Query a graph}
\usage{
querygraph(object, op, set = NULL, set2 = NULL, set3 = NULL)

ancestors(set, object)

ancestralSet(set, object)

parents(set, object)

children(set, object)

closure(set, object)

simplicialNodes(object)

ancestralGraph(set, object)

is.complete(object, set = NULL)

is.decomposition(set, set2, set3, object)

is.simplicial(set, object)
}
\arguments{
\item{object}{A graph}

\item{op}{The operation or query}

\item{set, set2, set3}{Sets of nodes in graph}
}
\description{
Unified approach to query a graph about its properties
}
