% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utilities.R
\name{ug2dag}
\alias{ug2dag}
\title{Coerce between undirected and directed graphs when possible}
\usage{
ug2dag(object)
}
\arguments{
\item{object}{An igraph object.}
}
\description{
An undirected graph G can be converted to a dag if G
is chordal.
}
