% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{gRbase-generics}
\alias{gRbase-generics}
\alias{fit}
\alias{compile}
\alias{propagate}
\alias{compile}
\alias{propagate}
\alias{stepwise}
\title{Compile and propagate functions}
\usage{
fit(object, ...)

compile(object, ...)

propagate(object, ...)

stepwise(object, ...)
}
\arguments{
\item{object}{An object}

\item{\dots}{Additional arguments which depends on the class of the object}
}
\value{
The value returned depends on the class of the first argument.
}
\description{
\code{compile} and \code{propagate} are generic
    functions which invoke particular methods which depend on the
    class of the first argument
}
\references{
Højsgaard, Søren; Edwards, David; Lauritzen, Steffen (2012):
    Graphical Models with R, Springer
}
\seealso{
\code{\link[gRain]{compile.grain}},
    \code{\link[gRain]{propagate.grain}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
