\name{gRbase-internal}

 \alias{subsetof}
 \alias{showg}
 \alias{readg}
 \alias{showf}
 \alias{readf}
 \alias{varset}
 \alias{in.list}
 \alias{is.cont}
 \alias{contains}

 \alias{remove.redundant}
 \alias{dual.rep}
 \alias{delete.edge}
 \alias{add.edge}

 \alias{is.graphical}


 \alias{out}
 \alias{ell}
 \alias{fun}
 \alias{ips}
 \alias{all.subsets}
 \alias{extract.power}

 \alias{corr.matrix}
 \alias{partial.corr.matrix}

% \alias{dynamic.Graph-methods}
% \alias{dynamic.Graph,hllm-method}
% \alias{dynamic.Graph,gmData-method}
 \alias{dynamic.Graph,gModel-method}
 \alias{dynamic.Graph,gRfit-method}
 \alias{label}
 \alias{width}
 \alias{label-methods}
% \alias{label,dg.Node-method}
 \alias{label,gModelTestClass-method}
% \alias{modifyModel-methods}
 \alias{modifyModel,gModel-method}
% \alias{testEdge-methods}
 \alias{testEdge,gModel-method}
 \alias{width-methods}
% \alias{width,dg.Edge-method}
 \alias{width,gModelTestClass-method}
 \alias{gModelTestClass-class}
% \alias{gModelTestClass}
 \alias{gRfit-class}

 \alias{dynamic.Graph}
 \alias{modifyModel}
% \alias{numberLevels}
% \alias{numberLevels<-}
 \alias{testEdge}

% \alias{all.subsets}
% \alias{extract.power}
 \alias{gREdges}
 \alias{dynamic.gR.Graph}
 \alias{gRVariableDescription}
 \alias{LabelAllEdges}
 \alias{selectOrder}
 \alias{UserMenus}

 \alias{processFormula}

 \alias{fit.ggm}
 \alias{ggm}
 \alias{stepwise}
 \alias{stepwise.default}



\title{Internal gRbase functions}
\description{
  Internal functions called by other functions.
}
\keyword{internal}
