\name{gRbase-internal}

%% Needs .Rd
\alias{coerceGraph}
\alias{coerceGraph.graphNEL}
\alias{coerceGraph.igraph}
\alias{coerceGraph.matrix}
\alias{isGSD_glist}
\alias{properties_glist}
\alias{isAcyclicMAT}
\alias{isUndirectedMAT}

%% Array operations in C.
\alias{aperm2}
\alias{cell2entry}
\alias{entry2cell}
\alias{factgrid}
\alias{getCellNumberC}
\alias{margcell2entry}
\alias{nextCell}
\alias{nextCellMarg}
\alias{permuteCellEntries}


%%%
%%% Functions comming from the gRain package
%%%

\alias{mcwhSH}
\alias{makeDimNames}
\alias{ripOrderGreen}
\alias{triangPG}
\alias{triangR}
\alias{triangRMAT}

\alias{igraph-class}

%%%
%%% SHD stuff
%%%

\alias{names2pairs}
\alias{rowmat2list}
\alias{colmat2list}
\alias{matrix2list}
\alias{ellK}

\alias{list2rhsFormula}
\alias{rhsFormula2list}
\alias{list2rhsf}
\alias{rhsf2list}

\alias{which.arr.ind}
\alias{pairs2num}
%% !!

%% DED stuff
\alias{showg}
\alias{readg}
\alias{showf}
\alias{readf}
\alias{varset}
\alias{in.list}
\alias{is.cont}
\alias{contains}
\alias{is.graphical}
\alias{dual.rep}
%% !!

%% ggm related stuff
\alias{fit.ggm}
\alias{ggm}
\alias{ips}
\alias{outfun}
\alias{ell}
%% !!

\alias{addEdge}
\alias{dropEdge}
\alias{addEdge.gModel}
\alias{dropEdge.gModel}



%%\alias{fun}

\alias{all.subsets}
\alias{extract.power}
\alias{selectOrder}
\alias{processFormula}
\alias{partial.corr.matrix}

\alias{update.gModel}
\alias{gRfit-class}

\alias{stepwise}
\alias{stepwise.default}



\title{Internal gRbase functions}
\description{
  Internal functions called by other functions.
}
\keyword{internal}

