\name{gRbase-internal}


%% SHD stuff
\alias{subsetof}
\alias{remove.redundant}
\alias{maximalSet}
\alias{isin}
\alias{names2pairs}
\alias{uniquePrim}
\alias{setdiffPrim}
\alias{unlistPrim}

%% !!

%% DED stuff
\alias{showg}
\alias{readg}
\alias{showf}
\alias{readf}
\alias{varset}
\alias{in.list}
\alias{is.cont}
\alias{contains}

\alias{is.graphical}
\alias{dual.rep}
\alias{delete.edge}
\alias{add.edge}
%% !!

%% ggm related stuff
\alias{fit.ggm}
\alias{ggm}
\alias{ips}
\alias{outfun}
\alias{ell}
%% !!


\alias{addEdge}
\alias{dropEdge}
\alias{addEdge.gModel}
\alias{dropEdge.gModel}



%%\alias{fun}

\alias{all.subsets}
\alias{extract.power}
\alias{selectOrder}
\alias{processFormula}



\alias{partial.corr.matrix}

\alias{update.gModel}
\alias{gRfit-class}

\alias{stepwise}
\alias{stepwise.default}



\title{Internal gRbase functions}
\description{
  Internal functions called by other functions.
}
\keyword{internal}

%%\alias{corr.matrix}
%%\alias{gRVariableDescription}
%%\alias{UserMenus}
%%\alias{gREdges}
%%\alias{dynamic.gR.Graph}
%%\alias{LabelAllEdges}

% \alias{dynamic.Graph}
% \alias{modifyModel}
% \alias{numberLevels}
% \alias{numberLevels<-}
% \alias{testEdge}

% \alias{all.subsets}
% \alias{extract.power}


 
% \alias{dynamic.Graph-methods}
% \alias{dynamic.Graph,hllm-method}
% \alias{dynamic.Graph,gmData-method}
% \alias{dynamic.Graph,gModel-method}
% \alias{dynamic.Graph,gRfit-method}
% \alias{label}
% \alias{width}
% \alias{label-methods}
% \alias{label,dg.Node-method}
% \alias{label,gModelTestClass-method}
% \alias{modifyModel-methods}
% \alias{modifyModel,gModel-method}
% \alias{testEdge-methods}
% \alias{testEdge,gModel-method}
% \alias{width-methods}
% \alias{width,dg.Edge-method}
% \alias{width,gModelTestClass-method}
% \alias{gModelTestClass-class}
% \alias{gModelTestClass}
