\name{gRbase-internal}

\alias{add.edge}
\alias{contains}
\alias{delete.edge}
\alias{dual.rep}
\alias{in.list}
\alias{is.cont}
\alias{is.graphical}
\alias{readf}
\alias{readg}
\alias{remove.redundant}
\alias{showf}
\alias{showg}
\alias{subsetof}
\alias{varset}

\alias{all.subsets}
\alias{extract.power}
\alias{gREdges}
\alias{dynamic.gR.Graph}
\alias{gRVariableDescription}
\alias{LabelAllEdges}
\alias{process.formula}
\alias{selectOrder}
\alias{UserMenus}

\title{gRbase internal functions}

\description{
  These are functions internally called by other functions
  in the package \pkg{gRbase} and not meant to be called by the user.
}

\keyword{internal}
