% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRbase-utilities.R
\name{gRbase-utilities}
\alias{gRbase-utilities}
\alias{rowmat2list}
\alias{matrix2list}
\alias{colmat2list}
\alias{which.arr.ind}
\alias{which_matrix_index}
\alias{rowSumsPrim}
\alias{colSumsPrim}
\alias{colwiseProd}
\alias{get_subset}
\alias{get_superset}
\alias{is_subsetof}
\title{gRbase utilities}
\usage{
rowmat2list(XX_)

colmat2list(XX_)

matrix2list(XX_, byrow = TRUE)

which.arr.ind(XX_)

which_matrix_index(XX_)

rowSumsPrim(XX_)

colSumsPrim(XX_)

colwiseProd(v, M)

get_subset(set, setlist, all = FALSE)

get_superset(set, setlist, all = FALSE)

is_subsetof(set, set2)
}
\arguments{
\item{XX_}{A matrix.}

\item{byrow}{Should the split be by row or by column.}

\item{v}{A vector.}

\item{M}{A matrix.}

\item{set, set2}{Character vectors}

\item{setlist}{List of charactervectors}

\item{all}{Logical.}
}
\description{
Various utility functions for gRbase. Includes 'faster
    versions' of certain standard R functions.
}
\details{
Returns matrix n x 2 matrix with indices of non-zero
    entries in matrix \code{XX_}. Notice
    \code{which_matrix_index__} is cpp implementation.

\code{colwiseProd} multiplies a vector v and a matrix M
    columnwise (as opposed to rowwise which is achieved by
    \code{v * M}). Hence \code{colwiseProd} does the same as
    \code{t(v * t(M))} - but it does so faster for numeric values.
}
\examples{
## colwiseProd
M <- matrix(1:16, nrow=4)
v <- 1:4
t(v * t(M))
colwiseProd(v, M)
system.time(for (ii in 1:100000)  t(v * t(M)))
system.time(for (ii in 1:100000)  colwiseProd(v, M))
}
