\name{gRbase}
\alias{gRbase}
\title{
The package `gRbase': summary information
}
\description{
  This package provides a basis for graphical modelling in R
}
\section{Details}{
\itemize{
  \item gRbase provides the general framework for setting up data and
  model structures and provide examples  for fitting hierarchical log linear
  models for contingency tables and graphical Gaussian models for the
  multivariate normal distribution.
  \item Other graphical model software on CRAN includes: mimR, deal,
  CoCo, ggm and SIN.
  }
The package is intended as a contribution to the gR-project 
described by Lauritzen (2002).
}
\section{Authors}{
Søren Højsgaard, Biometry Research Unit, Danish Institute of
Agricultural Sciences, DK-8830 Tjele, Denmark

Claus Dethlefsen, Center for Cardiovascular Research, Aalborg Hospital,
Århus University Hospital, DK-9000 Aalborg, Denmark
}
\section{Acknowledgements}{
  Thanks to the other members of the gR initiative, in particular to
  David Edwards for providing functions for formula-manipulation.
}
\references{
Lauritzen, S. L. (2002). gRaphical Models in R.
\emph{R News}, 3(2)39.
}
\keyword{multivariate}
\keyword{models}
\keyword{graphs}
