% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{internal}
\alias{internal}
\alias{matrix2dgCMatrix__}
\alias{dgCMatrix2matrix__}
\alias{M2dgCMatrix__}
\alias{M2matrix__}
\alias{internal}
\alias{dagList2dgCMatrix__}
\alias{ugList2dgCMatrix__}
\alias{dagList2matrix__}
\alias{ugList2matrix__}
\alias{adjList2tfList__}
\alias{adjList2ftList__}
\alias{adjList2tfM__}
\alias{adjList2ftM__}
\alias{adjList2matrix__}
\alias{adjList2dgCMatrix__}
\alias{internal}
\alias{which_matrix_index__}
\alias{rowmat2list__}
\alias{colmat2list__}
\alias{internal}
\alias{tab_list_mult_}
\alias{tab_list_add_}
\alias{tab_op_}
\alias{tab_add_}
\alias{tab_subt_}
\alias{tab_mult_}
\alias{tab_div_}
\alias{tab_div0_}
\alias{tab_equal_}
\alias{internal}
\alias{is_subsetof__}
\alias{get_superset__}
\alias{get_subset__}
\alias{internal}
\alias{all_pairs__}
\title{Internal Rcpp functions}
\description{
Automatically generated documentation of Rcpp functions.
}
