% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-grbase.R, R/yyyPrim.R
\name{internal}
\alias{internal}
\alias{ggm}
\alias{\%>\%}
\alias{MAT2ftM_}
\alias{isGSD_glist}
\alias{solveSPD}
\alias{randomGraph}
\alias{addEdge.gModel}
\alias{dropEdge.gModel}
\alias{dual.rep}
\alias{extract.power}
\alias{lapplyI2V}
\alias{lapplyV2I}
\alias{pairs2num}
\alias{processFormula}
\alias{selectOrder}
\alias{symMAT2ftM_}
\alias{setdiffPrim}
\alias{unlistPrim}
\alias{intersectPrim}
\alias{outerPrim}
\alias{uniquePrim}
\alias{matchPrim}
\title{Internal functions for the gRbase package}
\description{
These functions are not intended to be called directly.
}
