% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mathmark.R
\docType{data}
\name{mathmark}
\alias{math}
\alias{mathmark}
\title{Mathematics marks for students}
\format{This data frame contains the following columns: mechanics, vectors,
algebra, analysis, statistics.}
\usage{
mathmark
}
\description{
The \code{mathmark} data frame has 88 rows and 5 columns.
}
\examples{

data(mathmark)

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\references{
David Edwards, An Introduction to Graphical Modelling, Second
Edition, Springer Verlag, 2000
}
\keyword{datasets}

